+++ app/(gcforms)/[locale]/(form administration)/forms/actions.ts
@@ -5,6 +5,7 @@ import {
   deleteTemplate,
   getFullTemplateByID,
   cloneTemplate,
+  restoreTemplate,
 } from "@lib/templates";
 import { revalidatePath } from "next/cache";
 import { FormRecord } from "@lib/types";
@@ -13,9 +14,13 @@ import { AuthenticatedAction } from "@lib/actions";
 // Public facing functions - they can be used by anyone who finds the associated server action identifer
 
 export const getForm = AuthenticatedAction(
-  async (_, formId: string): Promise<{ formRecord: FormRecord | null; error?: string }> => {
+  async (
+    _,
+    formId: string,
+    allowDeleted: boolean
+  ): Promise<{ formRecord: FormRecord | null; error?: string }> => {
     try {
-      const response = await getFullTemplateByID(formId).catch(() => {
+      const response = await getFullTemplateByID(formId, allowDeleted).catch(() => {
         throw new Error("Failed to Get Form");
       });
       if (response === null) {
@@ -37,6 +42,22 @@ export const deleteForm = AuthenticatedAction(
           throw new Error("Responses Exist");
         } else {
           throw new Error("Failed to Delete Form");
+        }
+      });
+
+      revalidatePath("(gcforms)/[locale]/(form administration)/forms", "page");
+    } catch (e) {
+      return { error: (e as Error).message };
+    }
+  }
+);
+
+export const restoreForm = AuthenticatedAction(
+  async (_, id: string): Promise<void | { error?: string }> => {
+    try {
+      await restoreTemplate(id).catch((error) => {
+        if (error) {
+          throw new Error("Failed to Restore Form");
         }
       });