+++ packages/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [2.0.1] - 2025-11-12
+
+- Remove `.xml` and `.svg` file type support
+
+## [2.0.0] - 2025-11-05
+
+- Updates valueMatchesType to be more inline with `isFieldResponseValid` returns a t(string).
+
+- Adds valueMatches return <Boolean> use in place of `isFieldResponseValid` where a <Boolean> is needed.
 
 ## [1.0.11] - 2025-10-29